requirements.txt
requests
python-dotenv
openai
markdown 